Initial logic for recipe_handler set up
Constructors for Recipe and Instruction classes finished and tested.
Recipe now has a list of Instruction objects, and a counter that tracks which step it is currently on.
Each instruction object has attributes for Type, Value, and Instruction to Read
Next step is to set up logic for navigating steps